% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_random_helpers.R
\name{add_suffix}
\alias{add_suffix}
\title{Helper functions when creating get_all_missing()}
\usage{
add_suffix(df, suffix, keys = c())
}
\arguments{
\item{df}{a data frame}

\item{suffix}{a suffix to add}

\item{keys}{the joining keys}
}
\description{
Helper functions when creating get_all_missing(). Adds a suffix
}
