% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{baselines}
\alias{baselines}
\alias{statgains}
\title{Get Information on a Monster's Stat Baselines and Gains}
\usage{
baselines(main, sub = main)

statgains(main, sub = main)
}
\arguments{
\item{main}{The main type of the monster (character)}

\item{sub}{The sub type of the monster (character)}
}
\value{
A named numeric vector of six elements, one for each of the Lif,
Pow, Int, Ski, Spd, and Def stats
}
\description{
\code{baselines()} returns the monster's baseline stats.

\code{statgains()} returns the monster's stat gain information.
}
\examples{
baselines('Tiger')
baselines(main = 'Tiger', sub = 'Gali')
statgains('Tiger')
statgains(main = 'Tiger', sub = 'Gali')

}
