% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_table.R
\name{benchmark_table}
\alias{benchmark_table}
\title{Produce a table that returns a single test's value for each user selection position by percentile bin.}
\usage{
benchmark_table(
  test = "40yd",
  positions = c("DB", "DL", "LB", "OL", "QB", "RB", "TE", "WR", "PK", "LS")
)
}
\arguments{
\item{test}{character string. Start year of data gather (e.g. 2020). Data is only available starting from the year 2000.  If no start_year or end_year is specific, data from 2000-current is pulled.}

\item{positions}{character string. User defined positions to display percentiles for (e.g., 'DB', 'DL', 'LB', 'OL', 'QB', 'RB', 'TE', 'WR', 'PK', 'LS').}
}
\value{
A formatted table with combine values by percentile per position.
}
\description{
Produce a table that returns a single test's value for each user selection position by percentile bin.
}
\examples{
\dontrun{
benchmark_table (
test = '40yd',
positions = c('DB', 'DL', 'LB', 'OL', 'QB', 'RB', 'TE', 'WR')
)
}
}
