% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineR.r
\name{combineR}
\alias{combineR}
\title{combineR Package}
\description{
The \code{combineR} package includes three datasets and multiple functions to help you understand your data sets and na values from joins.
}
