% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_keys.R
\name{count_keys}
\alias{count_keys}
\title{See how many unique values of the joining key exist in your data.}
\usage{
count_keys(..., df_names = NULL, keys)
}
\arguments{
\item{...}{enter as many data frames as you would like}

\item{df_names}{the data frame names}

\item{keys}{the primary key or other key you wish to analyze}
}
\description{
Give this function as many data frames as you want and the key and it tells you how many unique values are in each data frame for the keys you specify.
}
\examples{
count_keys(demographics, math, ell, keys = c("student_id", "year"))
count_keys(list(demographics, math, ell),
     df_names = c("demographics", "math", "ell"),
     keys = c("student_id", "year"))

}
