% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{The \code{demographics} data}
\description{
The \code{demographics} data contains all students from an area in a time frame of 3 years. Columns are student id, year, name, sex, learning disability.
}
\keyword{data}
