% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ell.R
\docType{data}
\name{ell}
\alias{ell}
\title{The \code{ell} data}
\description{
The \code{ell} data shows ELL test. The ELL test is taken by the 40 percent flagged as potentially ELL students. Columns are student id, year, and score. Ell score is missing in 30 percent of cases. Scores are N(90 - 5 * learning disability, 15). Students no longer take the test once they score 100.
}
\keyword{data}
