% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_NAs.r
\name{explain_NAs}
\alias{explain_NAs}
\title{Given one dataset, find the number of NA values in each column for that dataset}
\usage{
explain_NAs(dataframe)
}
\arguments{
\item{dataframe}{Enter a dataset}
}
\description{
Given one dataset, find the number of NA values in each column for that dataset. Outputs a sentence, a data frame, and two plots
}
\examples{
explain_NAs(math)
explain_NAs(math)[[2]] # gives only the second part, which is the data frame
explain_NAs(math)[[3]] # gives only the third part, which is the first plot
}
