% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_rows.R
\name{explain_rows}
\alias{explain_rows}
\title{Find the number of rows, columns, and unique number of rows in your data sets}
\usage{
explain_rows(...)
}
\arguments{
\item{...}{enter as many data frames as you would like}
}
\description{
Enter as many data frames as you would like and find the number of rows, the number of columns, and the number of unique rows in your data sets
}
\examples{
explain_rows(demographics, math, ell)
}
