% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_missing.r
\name{get_all_missing}
\alias{get_all_missing}
\title{Get insight onto where missing values will enter when joining your data frames}
\usage{
get_all_missing(..., df_names = NULL, keys)
}
\arguments{
\item{...}{enter as many data frames as you would like}

\item{df_names}{the data frame names}

\item{keys}{the primary key in your data frames}
}
\description{
Enter data frames and a joining key and generate a table that shows you the variable, number of NA values in the data frames, and number of NA values introduced when full joining the data frames. Additionally, see this output in two different bar plots called 'Coded Missing Values Within Datasets' and 'Missing Values In Datasets Resulting From Joins'. This function enables you to specify as many data frames as you wish. You specify the keys, where each data frame must have at least one of the keys in it. The function gives insight into missing values and if they come from your current data or result from the join.
}
\examples{
get_all_missing(demographics, math, ell, keys = c("student_id", "year"))
get_all_missing(list(demographics, math, ell),
     df_names = c("demographics", "math", "ell"),
     keys = c("student_id", "year"))
}
