% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missings.R
\name{get_missings}
\alias{get_missings}
\title{Helper function used in get_all_missing()}
\usage{
get_missings(df, thename, the_count_keys, keys)
}
\arguments{
\item{df}{the data frame}

\item{thename}{the name of your data frame}

\item{the_count_keys}{counting keys}

\item{keys}{the joining keys}
}
\description{
Helper function used in get_all_missing()
}
