% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_random_helpers.R
\name{leftJoin}
\alias{leftJoin}
\title{Helper functions when creating get_all_missing()}
\usage{
leftJoin(x, y)
}
\arguments{
\item{x}{the first data frame}

\item{y}{the second data frame}
}
\description{
Helper functions when creating get_all_missing(). Conducts a left join
}
