% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monster.R
\name{Monster}
\alias{Monster}
\title{Create a 'Monster' class object}
\usage{
Monster(main, sub = main, stats = NULL)
}
\arguments{
\item{main}{The main type of the monster (character vector of length one)}

\item{sub}{The sub type of the monster (character vector of length one)}

\item{stats}{(Optional). The monster's six stats, Lif, Pow, Int, Ski, Spd,
and Def. If \code{NULL} (the default), the baseline baby stats for the monster
are used. (\emph{must} be a numeric vector of length six)}
}
\value{
A 'Monster' class object
}
\description{
Create a 'Monster' class object
}
\examples{
Monster(main = "Tiger", sub = "Gali")
Monster(main = "Mocchi", stats = c(300, 350, 350, 500, 400, 450))

}
