% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monsters}
\alias{monsters}
\title{Data on Monsters From Monster Rancher 2}
\format{
\subsection{\code{monsters}}{

A dataframe with 333 rows and 17 columns:
\describe{
\item{monster}{A unique identifier for the monster breeds, with the
format 'main/sub', where 'main' is replaced with the monster's main
type and 'sub' is replaced with the monster's sub type (see
\code{\link[=mr2types]{mr2types}})}
\item{main}{The main type of the monster (character)}
\item{sub}{The sub type of the monster (character)}
\item{lif_base}{The base 'Life' statistic for the monster at birth
(numeric)}
\item{pow_base}{The base 'Power' statistic for the monster at birth
(numeric)}
\item{int_base}{The base 'Intelligence' statistic for the monster at
birth (numeric)}
\item{ski_base}{The base 'Skill' statistic for the monster at birth
(numeric)}
\item{spd_base}{The base 'Speed' statistic for the monster at birth
(numeric)}
\item{def_base}{The base 'Defense' statistic for the monster at birth
(numeric)}
\item{lif_gain}{The stat gain for the monster, where 1 means very poor
gains and 5 means very high gains, which affect baby monster's stats
(see \code{vignette("breeding", package = 'combiner')}) (numeric)}
\item{pow_gain}{The stat gain for the monster (numeric)}
\item{int_gain}{The stat gain for the monster (numeric)}
\item{ski_gain}{The stat gain for the monster (numeric)}
\item{spd_gain}{The stat gain for the monster (numeric)}
\item{def_gain}{The stat gain for the monster (numeric)}
\item{tot_gain}{The total of the monster's stat gains (numeric)}
\item{lifespan}{The monster's base lifespan in weeks (character)}
}
}
}
\source{
This data is largely compiled from the GameFAQs Monster Rancher 2:
Walkthrough FAQ and Monster Rancher 2: Combining FAQ by Kurasu Soratobu,
available at \url{https://tinyurl.com/4s5677j9}.
}
\usage{
monsters
}
\description{
Certain attributes of monsters are relevant for determining combining
outcomes, specifically the parent monsters' stats, the 'baseline' stats for
the baby, and a 'stat gain' attribute that determines both how easily the
monster can improve that stat through training and how heavily that stat is
weighted in determining the combining outcome.
}
\seealso{
\code{\link[=mr2types]{mr2types}}
}
\keyword{datasets}
