% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mr2types}
\alias{mr2types}
\title{A Vector of all Valid Monster Rancher 2 Types}
\format{
\subsection{\code{mr2types}}{

A character vector giving all the valid monster types
}
}
\usage{
mr2types
}
\description{
When constructing a 'Monster' class object, you must provide a main and a
sub type for the monster. There are only so many types of monsters in
Monster Rancher 2, so this object provides the names of all the valid
monster types.
}
\details{
There are 38 monster types in Monster Rancher 2:
\itemize{
\item Ape
\item Arrow Head
\item Bajarl
\item Baku
\item Beaclon
\item Centaur
\item Color Pandora
\item Dragon
\item Ducken
\item Durahan
\item Gaboo
\item Gali
\item Ghost
\item Golem
\item Hare
\item Henger
\item Hopper
\item Jell
\item Jill
\item Joker
\item Kato
\item Metalner
\item Mew
\item Mocchi
\item Mock
\item Monol
\item Naga
\item Niton
\item Phoenix
\item Pixie
\item Plant
\item Suezo
\item Tiger
\item Undine
\item Worm
\item Wracky
\item Zilla
\item Zuum
}

Note that in this object, Arrow Head is recorded as "ArrowHead" and Color
Pandora is recorded as "ColorPandora".
}
\keyword{datasets}
