% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_combine_data.R
\name{pull_combine_data}
\alias{pull_combine_data}
\title{Pull NFL Combine Data from Pro Football Reference}
\usage{
pull_combine_data(start_year = 2000, end_year = 2022)
}
\arguments{
\item{start_year}{Year. Start year of data gather (e.g. 2020). Data is only available starting from the year 2000.  If no start_year or end_year is specific, data from 2000-current is pulled.}

\item{end_year}{Year. End year of data gather (e.g. 2021).}
}
\value{
A data frame with descriptive statistics. If you are only interested in certain columns
you can add these columns.
}
\description{
Pull NFL Combine Data from Pro Football Reference
}
\examples{
\dontrun{
pull_combine_data(
start_year = 2019,
end_year = 2021
)
}
}
