% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_random_helpers.R
\name{tally_missing}
\alias{tally_missing}
\title{Helper functions when creating get_all_missing()}
\usage{
tally_missing(df, keys)
}
\arguments{
\item{df}{a data frame}

\item{keys}{the joining keys}
}
\description{
Helper functions when creating get_all_missing(). Tallys the missing entries
}
